bina all masters

+++ Entity/SmartGrocery.Entity/DataModel/Society.cs
@@ -14,9 +14,9 @@ namespace SmartGrocery.Entity.DataModel
     
     public partial class Society
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
-        public Nullable<long> AreaId { get; set; }
+        public Nullable<int> AreaId { get; set; }
         public Nullable<bool> IsSelected { get; set; }
     
         public virtual Area Area { get; set; }
